Create main.py, etc.
Created main.py, initialized to a somewhat-imperfect but functional state based on some code I've been privately developing for a few days now. There's a lot of cleanup to do and in doing so I plan to split this off into its own files.

Currently the program is just a simple "test suite" of some basic functionalities for a NumPy-based simulation engine and graphics engine which I've accelerated via Numba.

Also created a .gitignore and added a list of required installs to the README.md