First PyGame App Demo
Achieved a ~50% FPS boost, from approx. 24 to 36, for 1000x1000px images.

Attempted to directly emulate the functionality of the App class in PyGameApp

Completely skirts the need to maintain and update GraphicsEngine.pilImage which helps a lot with processing speed.

Parallelization may be the next step to real improvement. Even though it's 3D and not specifically related, studying HackerPoet/PySpace may help as well.

Special thanks to Sentdex for his [Displaying Images with PyGame](https://pythonprogramming.net/displaying-images-pygame/) article

Also updated README.md to reference the above article and the PyGame documentation