Commit Before PyOpenGL Overhaul
Over the past few days I've been experimenting with OpenGL and PyOpenGL, namely with the goal of understanding Vertex Buffer Objects in order to take advantage of the GPU. I have managed to put together a simple rotating cuboid demo using VBOs for the vertices and triangles, but I did so in external sandbox files and "merging" will involve a lot of deletions and overwrites. I want to ensure that I have a record of these changes.