Intermediate Commit for Rotation (save state)
NOTE: Working on rotation matrix math, with the goal of eventually implementing numpy-accelerated rotation on a per-shape basis. Just managed to fix enough bugs that the basic display functionality is working properly, and this commit serves to save the current code state and its differences relative to the previous commit.

Includes random seeding.