Fix Incorrect Print Statements in nsvt_geometry.py
In developing a separate private project which makes use of the same rotation methods as NSVT, I noticed that some of the invariant-related print statements were printing incorrect statements related to the names of methods during invariant check failure.

These are extremely minor errors, but because this is a public project which I may not return to for some time, I find it important to update the repository immediately.